Add systemd support
This compiles postgres w/ systemd support and adds 2 unit files, one
for handling the initial configuration (similar to what mariadb does),
and one for actually running postgres.

The initial postgres database is created via the postgresql-install.service.
The default database is created in /var/lib/pgsql/data. The install service
runs when /var/lib/pgsql/ does not exist. The initial configuration rejects all
local and remote connections, so the admin has to manually configure access
in /var/lib/pgsql/pg_hba.conf before real use.

The actual postgres server is controlled via postgresql.service. This file is
adapted from the example service file in the Postgres documentation:
https://www.postgresql.org/docs/devel/static/server-start.html

The postgres server and it's utilities, e.g., initdb, are required to be
run as the postgres user. As a result, this change depends on the patch
to clr-systemd-config that adds the postgres user.